requirements.txt
python-crfsuite
sklearn-crfsuite
joblib
tensorflow
gensim
keras
pandas
numpy
django
djangorestframework
django-cors-headers